Implement rough draft of StreamWriter, share code with FileWriter
